-Updated simplifiedVariant coordinates from int to long -Updated start and end from int to long in different services parameters

+++ variation-commons-mongodb/src/main/java/uk/ac/ebi/eva/commons/mongodb/entities/projections/SimplifiedVariant.java
@@ -53,10 +53,10 @@ public class SimplifiedVariant {
     private String chromosome;
 
     @Field(START_FIELD)
-    private int start;
+    private long start;
 
     @Field(END_FIELD)
-    private int end;
+    private long end;
 
     @Field(LENGTH_FIELD)
     private int length;
@@ -77,7 +77,7 @@ public class SimplifiedVariant {
         this(null, null, -1, -1, -1, null, null, null);
     }
 
-    public SimplifiedVariant(VariantType variantType, String chromosome, int start, int end, int length,
+    public SimplifiedVariant(VariantType variantType, String chromosome, long start, long end, int length,
                              String reference, String alternate, Map<String, Set<String>> hgvs) {
         this.id = buildVariantId(chromosome, start, reference, alternate);
         this.variantType = variantType;
